[Big work] + [dont forget to put functionality to the services from controllers]

+++ src/main/java/marks/webstore/controller/ProductReviewController.java
@@ -2,20 +2,20 @@ package marks.webstore.controller;
 
 import marks.webstore.domain.ProductType;
 import marks.webstore.domain.ProductTypeStore;
-import marks.webstore.repos.ProductTypeStoreRepo;
-import marks.webstore.service.ProductStoreService;
-import org.springframework.beans.factory.annotation.Autowired;
-import marks.webstore.repos.ProductTypeRepo;
-import marks.webstore.repos.ProductTypeStoreRepo;
-import marks.webstore.repos.StoreRepo;
+import marks.webstore.domain.User;
+import marks.webstore.repos.UserStoreRepo;
 import marks.webstore.service.ProductService;
 import marks.webstore.service.ProductStoreService;
 import marks.webstore.service.StoreService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.Map;
+import java.util.stream.Collectors;
 
 @Controller
 @RequestMapping("/products")
@@ -30,8 +30,15 @@ public class ProductReviewController {
     @Autowired
     private ProductService productService;
 
+    @Autowired
+    private UserStoreRepo userStoreRepo;
+
     @GetMapping("{product}")
     public String productReviewForm(@PathVariable ProductType product, Model model) {
+        if (productStoreService.findStoreByProduct(product) == null ||
+        !product.getPublished()) {
+            return "resources/public/error/404";
+        }
         ProductTypeStore productTypeStore = productStoreService.findStoreByProduct(product);
 
         model.addAttribute("store", productTypeStore);
@@ -41,8 +48,16 @@ public class ProductReviewController {
 
     @PreAuthorize("hasAnyAuthority('ADMIN', 'REDACTOR')")
     @GetMapping("{product}/edit")
-    public String editProduct(@PathVariable ProductType product, Model model) {
-        Long amountOfProducts = productStoreService.findAllByProductId(product.getId()).get(0).getAmount();
+    public String editProduct(@PathVariable ProductType product,
+                              Model model,
+                              @AuthenticationPrincipal User user) {
+        if (productStoreService.findStoreByProduct(product) == null) {
+            return "resources/public/error/404";
+        }
+        if (!productStoreService.isRedactorHasRoot(user, product)) {
+            return "resources/public/error/403";
+        }
+        Long amountOfProducts = productStoreService.getAmountOfProducts(product);
         model.addAttribute("product", product);
         model.addAttribute("stores", storeService.findAllStores());
         model.addAttribute("amountofproducts", amountOfProducts);
@@ -54,21 +69,41 @@ public class ProductReviewController {
     public String updateProduct(
             @PathVariable ProductType product,
             @RequestParam String name,
-            @RequestParam String storeName,
-            @RequestParam Float price,
+            String storeName,
+            @RequestParam Double price,
             @RequestParam Long amount,
-            @RequestParam String description
-            ) {
-        productService.updateProduct(product, name, storeName, price, amount, description);
+            @RequestParam String description,
+            @AuthenticationPrincipal User user
+    ) {
+        if (productStoreService.findStoreByProduct(product) == null) {
+            return "resources/public/error/404";
+        }
+        if (!productStoreService.isRedactorHasRoot(user, product)) {
+            return "resources/public/error/403";
+        }
+        productService.updateProduct(product, name, storeName, price, amount, description, user.isAdmin());
 
-        return "redirect:/products/{product}";
+        return "redirect:/productsList";
     }
 
     @PreAuthorize("hasAnyAuthority('ADMIN', 'REDACTOR')")
     @GetMapping("{product}/delete")
     public String deleteProduct(
-            @PathVariable ProductType product) {
+            @AuthenticationPrincipal User user,
+            @PathVariable ProductType product,
+            Map<String, Object> model) {
+        if (productStoreService.findStoreByProduct(product) == null) {
+            return "resources/public/error/404";
+        }
+        if (!productStoreService.isRedactorHasRoot(user, product)) {
+            return "resources/public/error/403";
+        }
         productService.deleteProduct(product);
-        return "redirect:/products";
+        model.put("producttypes", user.isAdmin() ?  productService.findAllProductsReverse() :
+                productService.findAllProductsReverse().stream()
+                        .filter(productType -> productStoreService.isRedactorHasRoot(user, productType))
+                        .collect(Collectors.toList()));
+
+        return "redactorProductList";
     }
 }
